nrf/ppi: Fix add/remove channel for DPPI.
CHG[n] ignores writes when EN/DIS subscriptions are enabled; temporarily
disable subscriptions to write CHG[n].
