+++ embassy-nrf/src/ppi/mod.rs
@@ -56,32 +56,6 @@ impl<'d, G: Group> PpiGroup<'d, G> {
         Self { g }
     }
 
-    /// Add a PPI channel to this group.
-    ///
-    /// If the channel is already in the group, this is a no-op.
-    pub fn add_channel<C: Channel, const EVENT_COUNT: usize, const TASK_COUNT: usize>(
-        &mut self,
-        ch: &Ppi<'_, C, EVENT_COUNT, TASK_COUNT>,
-    ) {
-        let r = self.g.regs();
-        let ng = self.g.number();
-        let nc = ch.ch.number();
-        r.chg(ng).modify(|w| w.set_ch(nc, true));
-    }
-
-    /// Remove a PPI channel from this group.
-    ///
-    /// If the channel is already not in the group, this is a no-op.
-    pub fn remove_channel<C: Channel, const EVENT_COUNT: usize, const TASK_COUNT: usize>(
-        &mut self,
-        ch: &Ppi<'_, C, EVENT_COUNT, TASK_COUNT>,
-    ) {
-        let r = self.g.regs();
-        let ng = self.g.number();
-        let nc = ch.ch.number();
-        r.chg(ng).modify(|w| w.set_ch(nc, false));
-    }
-
     /// Enable all the channels in this group.
     pub fn enable_all(&mut self) {
         let n = self.g.number();
